pyproject.toml
[tool.poetry]
name = "Embedding"
version = "0.0.1"
description = "RDB to VectorDB"
authors = ["jaeyulim <twilight4j@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

# 언어 및 자연어 처리 관련 패키지
langchain = "0.*"
langchain-core = "0.*"
langchain-experimental = "0.*"
langchain-community = "0.*"
langchain-openai = "0.*"
langchain-elasticsearch = "0.*"
langchain-chroma = "0.*"
langchainhub = "0.*"
langgraph = "0.*"
langsmith = "0.*"
openai = "1.*"


# 데이터베이스 및 캐시 관련 패키지
redis = "5.*"
chromadb = "0.*"

# 시각화 및 사용자 인터페이스 관련 패키지
matplotlib = "3.*"
streamlit = "1.*"
jupyter = "1.*"
notebook = "7.*"

# 딥러닝 및 머신러닝 관련 패키지
faiss-cpu = "1.*"

# 기타 유틸리티 및 필수 패키지
python-dotenv = "1.*"
oracledb = "3.*"
protobuf = "3.*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
